fix: classify transport errors by type instead of string matching
- Use hms_client_error helper to preserve ClientError::Transport
  classification when converting to CatalogError
- Replace all Err(err) => CatalogError::External sites with typed helper
- Fix clippy: remove unnecessary casts, use to_vec() instead of
  iter().cloned().collect(), allow expect_used in test
- should_retry now matches 'transport error' prefix from typed helper

+++ crates/sail-catalog-hms/src/data_type.rs
@@ -21,8 +21,8 @@ pub fn arrow_to_hive_type(data_type: &DataType) -> CatalogResult<String> {
         | DataType::Decimal64(precision, scale)
         | DataType::Decimal128(precision, scale)
         | DataType::Decimal256(precision, scale) => {
-            let precision = *precision as u8;
-            let scale = *scale as i8;
+            let precision = *precision;
+            let scale = *scale;
             if precision > 38 {
                 return Err(CatalogError::InvalidArgument(format!(
                     "Hive Metastore supports decimal precision up to 38, got {precision}"
